+++ src/Telegram.Bot/Types/InputFiles/IInputMedia.cs
@@ -1,8 +1,11 @@
-﻿using Telegram.Bot.Types.Enums;
+using Telegram.Bot.Types.Enums;
 
 // ReSharper disable once CheckNamespace
 namespace Telegram.Bot.Types
 {
+    /// <summary>
+    /// Media to send in request that could be a file_id, HTTP url, or a file
+    /// </summary>
     public interface IInputMedia
     {
         /// <summary>